refactor: by comments/ add children to Header and fix form-login styles
