--- project/src/mocks/offers.ts
@@ -1,154 +1,0 @@
-export const offers = [
-  {
-    'bedrooms': 3,
-    'city': {
-      'location': {
-        'latitude': 52.370216,
-        'longitude': 4.895168,
-        'zoom': 10
-      },
-      'name': 'Amsterdam'
-    },
-    'description': 'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
-    'goods': [
-      'Heating', 'Wi-Fi', 'Washing machine', 'Towels', 'Coffee machine', 'Baby seat', 'Kitchen', 'Dishwasher', 'Cabel TV', 'Fridge',
-    ],
-    'host': {
-      'avatarUrl': 'img/avatar-angelina.jpg',
-      'id': 3,
-      'isPro': true,
-      'name': 'Angelina'
-    },
-    'id': 1,
-    'images': [
-      'img/room.jpg', 'img/apartment-01.jpg', 'img/apartment-02.jpg', 'img/apartment-03.jpg',
-    ],
-    'isFavorite': true,
-    'isPremium': false,
-    'location': {
-      'latitude': 52.3909553943508,
-      'longitude': 4.85309666406198,
-      'zoom': 8
-    },
-    'maxAdults': 4,
-    'previewImage': 'img/room.jpg',
-    'price': 120,
-    'rating': 4.8,
-    'title': 'Beautiful & luxurious studio at great location',
-    'type': 'apartment'
-  },
-  {
-    'bedrooms': 1,
-    'city': {
-      'location': {
-        'latitude': 52.370216,
-        'longitude': 4.895168,
-        'zoom': 10
-      },
-      'name': 'Amsterdam'
-    },
-    'description': 'Lorem ipsum dolor sit amet consectetur adipisicing elit.',
-    'goods': [
-      'Wi-Fi', 'Towels', 'Cabel TV', 'Fridge',
-    ],
-    'host': {
-      'avatarUrl': 'img/avatar-max.jpg',
-      'id': 4,
-      'isPro': false,
-      'name': 'Truffaldino'
-    },
-    'id': 2,
-    'images': [
-      'img/room.jpg', 'img/apartment-01.jpg', 'img/apartment-02.jpg', 'img/apartment-03.jpg',
-    ],
-    'isFavorite': true,
-    'isPremium': true,
-    'location': {
-      'latitude': 52.369553943508,
-      'longitude': 4.85309666406198,
-      'zoom': 8
-    },
-    'maxAdults': 2,
-    'previewImage': 'img/room.jpg',
-    'price': 100,
-    'rating': 5.0,
-    'title': 'Studio channel view',
-    'type': 'flat'
-  },
-  {
-    'bedrooms': 8,
-    'city': {
-      'location': {
-        'latitude': 52.370216,
-        'longitude': 4.895168,
-        'zoom': 10
-      },
-      'name': 'Amsterdam'
-    },
-    'description': 'Atque sapiente magni ex consectetur quos optio suscipit, repellendus dolorum culpa laborum sint nobis facilis debitis ea minima tenetur minus voluptate dicta!',
-    'goods': [
-      'Heating', 'Wi-Fi', 'Washing machine', 'Towels', 'Coffee machine', 'Baby seat', 'Kitchen', 'Dishwasher', 'Cabel TV', 'Fridge',
-    ],
-    'host': {
-      'avatarUrl': 'img/avatar-angelina.jpg',
-      'id': 5,
-      'isPro': true,
-      'name': 'Luigina'
-    },
-    'id': 3,
-    'images': [
-      'img/room.jpg', 'img/apartment-01.jpg', 'img/apartment-02.jpg', 'img/apartment-03.jpg',
-    ],
-    'isFavorite': false,
-    'isPremium': true,
-    'location': {
-      'latitude': 52.3909553943508,
-      'longitude': 4.929309666406198,
-      'zoom': 8
-    },
-    'maxAdults': 20,
-    'previewImage': 'img/room.jpg',
-    'price': 150,
-    'rating': 3.9,
-    'title': 'Superhotel',
-    'type': 'hotel'
-  },
-  {
-    'bedrooms': 2,
-    'city': {
-      'location': {
-        'latitude': 52.370216,
-        'longitude': 4.895168,
-        'zoom': 10
-      },
-      'name': 'Amsterdam'
-    },
-    'description': 'Eum et veniam dolore debitis quos corrupti, facilis saepe laboriosam iure, laborum modi!',
-    'goods': [
-      'Wi-Fi', 'Towels', 'Coffee machine', 'Kitchen', 'Cabel TV', 'Fridge',
-    ],
-    'host': {
-      'avatarUrl': 'img/avatar-max.jpg',
-      'id': 6,
-      'isPro': true,
-      'name': 'Paul'
-    },
-    'id': 4,
-    'images': [
-      'img/room.jpg', 'img/apartment-01.jpg', 'img/apartment-02.jpg', 'img/apartment-03.jpg',
-    ],
-    'isFavorite': true,
-    'isPremium': false,
-    'location': {
-      'latitude': 52.3809553943508,
-      'longitude': 4.939309666406198,
-      'zoom': 8
-    },
-    'maxAdults': 3,
-    'previewImage': 'img/room.jpg',
-    'price': 80,
-    'rating': 4.5,
-    'title': 'Paul and Anna',
-    'type': 'apartment'
-  }
-];
