--- project/src/mocks/reviews.ts
@@ -1,26 +1,0 @@
-export const reviews = [
-  {
-    'comment': 'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
-    'date': 'Fri Jul 15 2022 13:20:24 GMT+0700 (Novosibirsk Standard Time)',
-    'id': 10,
-    'rating': 4,
-    'user': {
-      'avatarUrl': 'img/avatar-max.jpg',
-      'id': 1,
-      'isPro': false,
-      'name': 'Max'
-    }
-  },
-  {
-    'comment': 'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
-    'date': 'Mon Jul 19 2022 13:20:24 GMT+0700 (Novosibirsk Standard Time)',
-    'id': 20,
-    'rating': 5,
-    'user': {
-      'avatarUrl': 'img/avatar-angelina.jpg',
-      'id': 2,
-      'isPro': true,
-      'name': 'Angelina'
-    }
-  },
-];
