--- project/src/mocks/user.ts
@@ -1,8 +1,0 @@
-export const user = {
-  'avatarUrl': 'img/1.png',
-  'email': 'Oliver.conner@gmail.com',
-  'id': 1,
-  'isPro': false,
-  'name': 'Oliver.conner',
-  'token': 'T2xpdmVyLmNvbm5lckBnbWFpbC5jb20=',
-};
